Add handling of multiple arguments
When Slack sends the POST, the argument(s) of the slash command are sent
in the `text` field. If there is a space (e.g. multiple arguments)
this is encoded using the `+` sign.
To create the raw body request of the POST, the code is using the
`querystring.stringify()`. This encodes the space as `%20`.
Workaround for that is to replace `%20` with `+` sign.

Fix #1